backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
torch==2.1.0
torchvision==0.16.0
timm==0.9.12
numpy>=1.24.0,<2.0.0
pillow>=10.0.0
pydantic==2.5.0
google-cloud-storage==2.10.0
